[blobserve]: export the config as a package
